Fix issues identified by mypy checks
